Further refactor to separate concerns further, prevent circ dependencies
