+++ server/models/services/AtService.js
@@ -4,24 +4,10 @@ const {
   AT_VERSION_ATTRIBUTES,
   BROWSER_ATTRIBUTES
 } = require('./helpers');
-const { Sequelize, At, AtVersion } = require('../');
+const { Sequelize, At } = require('../');
 const { Op } = Sequelize;
-const { TestPlanReport } = require('../');
-const {
-  cloneTestPlanReportWithNewAtVersion
-} = require('./TestPlanReportService');
-const { getAtVersions } = require('./AtVersionService');
 
 // association helpers to be included with Models' results
-/**
- * @param atAttributes - At attributes
- * @returns {{association: string, attributes: string[]}}
- */
-const atAssociation = atAttributes => ({
-  association: 'at',
-  attributes: atAttributes
-});
-
 /**
  * @param atVersionAttributes - AtVersion attributes
  * @returns {{association: string, attributes: string[]}}
@@ -188,355 +174,11 @@ const removeAtById = async ({ id, truncate = false, transaction }) => {
   return ModelService.removeById(At, { id, truncate, transaction });
 };
 
-// AtVersion
-
-/**
- * You can pass any of the attribute arrays as '[]' to exclude that related association
- * @param {number} options.id - unique id of the AtVersion model being queried
- * @param {object} options
- * @param {string[]} options.atVersionAttributes  - AtVersion attributes to be returned in the result
- * @param {string[]} options.atAttributes  - At attributes to be returned in the result
- * @param {*} options.transaction - Sequelize transaction
- * @returns {Promise<*>}
- */
-const getAtVersionById = async ({
-  id,
-  atVersionAttributes = AT_VERSION_ATTRIBUTES,
-  atAttributes = AT_ATTRIBUTES,
-  transaction
-}) => {
-  return ModelService.getById(AtVersion, {
-    id,
-    attributes: atVersionAttributes,
-    include: [atAssociation(atAttributes)],
-    transaction
-  });
-};
-
-/**
- * You can pass any of the attribute arrays as '[]' to exclude that related association
- * @param {object} options
- * @param {object} options.where - unique values of the AtVersion model being queried
- * @param {string[]} options.atVersionAttributes  - AtVersion attributes to be returned in the result
- * @param {string[]} options.atAttributes  - At attributes to be returned in the result
- * @param {*} options.transaction - Sequelize transaction
- * @returns {Promise<*>}
- */
-const getAtVersionByQuery = async ({
-  where: { atId, name, releasedAt },
-  atVersionAttributes = AT_VERSION_ATTRIBUTES,
-  atAttributes = AT_ATTRIBUTES,
-  transaction
-}) => {
-  return ModelService.getByQuery(AtVersion, {
-    where: {
-      atId,
-      ...(name && { name }),
-      ...(releasedAt && { releasedAt })
-    },
-    attributes: atVersionAttributes,
-    include: [atAssociation(atAttributes)],
-    transaction
-  });
-};
-
-/**
- * @param {object} options
- * @param {object} options.createParams - values to be used to create the AtVersion record
- * @param {string[]} options.atVersionAttributes  - AtVersion attributes to be returned in the result
- * @param {string[]} options.atAttributes  - At attributes to be returned in the result
- * @param {*} options.transaction - Sequelize transaction
- * @returns {Promise<*>}
- */
-const createAtVersion = async ({
-  values: { atId, name, releasedAt },
-  atVersionAttributes = AT_VERSION_ATTRIBUTES,
-  atAttributes = AT_ATTRIBUTES,
-  transaction
-}) => {
-  await ModelService.create(AtVersion, {
-    values: { atId, name, releasedAt },
-    transaction
-  });
-  // to ensure the structure being returned matches what we expect for simple queries and can be controlled
-  return ModelService.getByQuery(AtVersion, {
-    where: { atId, name },
-    attributes: atVersionAttributes,
-    include: [atAssociation(atAttributes)],
-    transaction
-  });
-};
-
-/**
- * @param {object} options
- * @param {object} options.where - values of the AtVersion record to be updated
- * @param {object} options.values - values to be used to update columns for the record being referenced
- * @param {string[]} options.atVersionAttributes  - AtVersion attributes to be returned in the result
- * @param {string[]} options.atAttributes  - At attributes to be returned in the result
- * @param {*} options.transaction - Sequelize transaction
- * @returns {Promise<*>}
- */
-const updateAtVersionByQuery = async ({
-  where: { atId, name, releasedAt },
-  values = {},
-  atVersionAttributes = AT_VERSION_ATTRIBUTES,
-  atAttributes = AT_ATTRIBUTES,
-  transaction
-}) => {
-  await ModelService.update(AtVersion, {
-    where: { atId, name, releasedAt },
-    values,
-    transaction
-  });
-  return ModelService.getByQuery(AtVersion, {
-    where: { atId, name: values.name || name, releasedAt },
-    attributes: atVersionAttributes,
-    include: [atAssociation(atAttributes)],
-    transaction
-  });
-};
-
-/**
- * @param {object} options
- * @param {number} options.id - id of the AtVersion record to be updated
- * @param {object} options.values - values to be used to update columns for the record being referenced
- * @param {string[]} options.atVersionAttributes  - AtVersion attributes to be returned in the result
- * @param {string[]} options.atAttributes  - At attributes to be returned in the result
- * @param {*} options.transaction - Sequelize transaction
- * @returns {Promise<*>}
- */
-const updateAtVersionById = async ({
-  id,
-  values = {},
-  atVersionAttributes = AT_VERSION_ATTRIBUTES,
-  atAttributes = AT_ATTRIBUTES,
-  transaction
-}) => {
-  await ModelService.update(AtVersion, {
-    where: { id },
-    values,
-    transaction
-  });
-  return ModelService.getById(AtVersion, {
-    id,
-    attributes: atVersionAttributes,
-    include: [atAssociation(atAttributes)],
-    transaction
-  });
-};
-
-/**
- * @param {object} options
- * @param {object} options.where - values of the AtVersion record to be removed
- * @param {boolean} options.truncate - Sequelize specific deletion options that could be passed
- * @param {*} options.transaction - Sequelize transaction
- * @returns {Promise<boolean>}
- */
-const removeAtVersionByQuery = async ({
-  where: { atId, name, releasedAt },
-  truncate = false,
-  transaction
-}) => {
-  return ModelService.removeByQuery(AtVersion, {
-    where: { atId, name, releasedAt },
-    truncate,
-    transaction
-  });
-};
-
-/**
- * @param {object} options
- * @param {number} options.id - id of the AtVersion record to be removed
- * @param {boolean} options.truncate - Sequelize specific deletion options that could be passed
- * @param {*} options.transaction - Sequelize transaction
- * @returns {Promise<boolean>}
- */
-const removeAtVersionById = async ({ id, truncate = false, transaction }) => {
-  return ModelService.removeById(AtVersion, {
-    id,
-    truncate,
-    transaction
-  });
-};
-
-/**
- * Returns all the unique AT Versions used when collecting results from testers
- * for a Test Plan Report
- * @param {number} testPlanReportId - id of the test plan report
- * @param {object} options
- * @param {*} options.transaction - Sequelize transaction
- * @returns {Promise<*>}
- */
-const getUniqueAtVersionsForReport = async (
-  testPlanReportId,
-  { transaction }
-) => {
-  const results = await ModelService.rawQuery(
-    `
-      select "At".name        as "atName",
-             "atVersionId",
-             "AtVersion".name as "atVersionName",
-             "releasedAt",
-             "testPlanReportId",
-             "testerUserId",
-             "testPlanRunId"
-      from ( select distinct "TestPlanReport".id                                              as "testPlanReportId",
-                             "TestPlanRun".id                                                 as "testPlanRunId",
-                             "TestPlanRun"."testerUserId",
-                             (jsonb_array_elements("testResults") ->> 'atVersionId')::integer as "atVersionId"
-             from "TestPlanReport"
-                    left outer join "TestPlanRun" on "TestPlanRun"."testPlanReportId" = "TestPlanReport".id
-             where "testPlanReportId" = ${testPlanReportId}
-             group by "TestPlanReport".id, "TestPlanRun".id ) as atVersionResults
-             join "AtVersion" on "AtVersion".id = atVersionResults."atVersionId"
-             join "At" on "AtVersion"."atId" = "At".id;
-    `,
-    { transaction }
-  );
-
-  // Sort in descending order of releasedAt date
-  results.sort((a, b) => new Date(b.releasedAt) - new Date(a.releasedAt));
-
-  return results;
-};
-
-/**
- * @param {object} options
- * @param {object} options.where - values to be used to create or find the AtVersion record
- * @param {string[]} options.atVersionAttributes  - AtVersion attributes to be returned in the result
- * @param {string[]} options.atAttributes  - At attributes to be returned in the result
- * @param {*} options.transaction - Sequelize transaction
- * @returns {BrowserVersion}
- */
-const findOrCreateAtVersion = async ({
-  where: { atId, name, releasedAt },
-  atVersionAttributes = AT_VERSION_ATTRIBUTES,
-  atAttributes = AT_ATTRIBUTES,
-  transaction
-}) => {
-  let version = await getAtVersionByQuery({
-    where: { atId, name },
-    atVersionAttributes,
-    atAttributes,
-    transaction
-  });
-
-  if (!version) {
-    /* TODO: releasedAt manually entered by users submitting new version, support actual dates in automation */
-    version = await createAtVersion({
-      values: { atId, name, releasedAt: releasedAt ?? new Date() },
-      atVersionAttributes,
-      atAttributes,
-      transaction
-    });
-  }
-
-  return version;
-};
-
-/**
- * Helper function to schedule a collection job for a test plan report
- * @param {object} options
- * @param {number} options.testPlanReportId - ID of the test plan report
- * @param {*} options.transaction - Sequelize transaction
- * @returns {Promise<*>} The created collection job
- */
-const scheduleJob = async ({ testPlanReportId, transaction }) => {
-  const { scheduleCollectionJob } = require('./CollectionJobService');
-  return scheduleCollectionJob({ testPlanReportId }, { transaction });
-};
-
-/**
- * Gets the most recent previous AT version for a given AT version and creates collection jobs
- * for all finalized test plan runs that used that previous version
- * @param {object} options
- * @param {number} options.atVersionId - ID of the current AT version
- * @param {*} options.transaction - Sequelize transaction
- * @returns {Promise<Array>} - Array of created collection jobs
- */
-const createCollectionJobsFromPreviousVersion = async ({
-  atVersionId,
-  transaction
-}) => {
-  // Get the current AT version with its AT info
-  const currentVersion = await getAtVersionById({
-    id: atVersionId,
-    transaction
-  });
-
-  if (!currentVersion) {
-    throw new Error(`AT Version with ID ${atVersionId} not found`);
-  }
-
-  // Find the most recent previous version for the same AT
-  const previousVersion = await AtVersion.findOne({
-    where: {
-      atId: currentVersion.atId,
-      releasedAt: {
-        [Op.lt]: currentVersion.releasedAt
-      }
-    },
-    order: [['releasedAt', 'DESC']],
-    transaction
-  });
-
-  if (!previousVersion) {
-    throw new Error(
-      `No previous version found for AT Version ID ${atVersionId}`
-    );
-  }
-
-  // Get all finalized test plan reports that used the previous version
-  const finalizedReports = await TestPlanReport.findAll({
-    where: {
-      [Op.or]: [
-        { exactAtVersionId: previousVersion.id },
-        { minimumAtVersionId: previousVersion.id }
-      ],
-      markedFinalAt: {
-        [Op.not]: null
-      }
-    },
-    transaction
-  });
-
-  const collectionJobs = [];
-  for (const report of finalizedReports) {
-    // Clone the historical report with the new current AT version
-    const newReport = await cloneTestPlanReportWithNewAtVersion(
-      report,
-      currentVersion,
-      transaction
-    );
-    const job = await scheduleJob({
-      testPlanReportId: newReport.id,
-      transaction
-    });
-    collectionJobs.push(job);
-  }
-  return collectionJobs;
-};
-
 module.exports = {
   // Basic CRUD [At]
   getAtById,
   getAts,
   createAt,
   updateAtById,
-  removeAtById,
-
-  // Basic CRUD [AtVersion]
-  getAtVersionById,
-  getAtVersionByQuery,
-  getAtVersions,
-  createAtVersion,
-  updateAtVersionById,
-  updateAtVersionByQuery,
-  removeAtVersionByQuery,
-  removeAtVersionById,
-
-  // Custom Methods
-  getUniqueAtVersionsForReport,
-  findOrCreateAtVersion,
-  createCollectionJobsFromPreviousVersion
+  removeAtById
 };
