+++ server/resolvers/AtOperations/findOrCreateAtVersionResolver.js
@@ -1,5 +1,7 @@
 const { AuthenticationError } = require('apollo-server');
-const { findOrCreateAtVersion } = require('../../models/services/AtService');
+const {
+  findOrCreateAtVersion
+} = require('../../models/services/AtVersionService');
 
 const findOrCreateAtVersionResolver = async (
   { parentContext: { id: atId } },
