+++ server/resolvers/AtVersionOperations/updateAtVersionResolver.js
@@ -1,5 +1,7 @@
 const { AuthenticationError } = require('apollo-server');
-const { updateAtVersionById } = require('../../models/services/AtService');
+const {
+  updateAtVersionById
+} = require('../../models/services/AtVersionService');
 
 const updateAtVersionResolver = async (
   { parentContext: { id } },
