+++ server/resolvers/index.js
@@ -50,7 +50,7 @@ const TestPlanRun = require('./TestPlanRun');
 const Test = require('./Test');
 const ScenarioResult = require('./ScenarioResult');
 const AtVersion = require('./AtVersion');
-const createCollectionJobsFromPreviousVersion = require('./AtOperations/createCollectionJobsFromPreviousVersionResolver');
+const createCollectionJobsFromPreviousVersion = require('./createCollectionJobsFromPreviousVersionResolver');
 
 const resolvers = {
   Query: {
