+++ server/resolvers/scheduleCollectionJobResolver.js
@@ -2,9 +2,15 @@ const { AuthenticationError } = require('apollo-server-core');
 const {
   scheduleCollectionJob
 } = require('../models/services/CollectionJobService');
+const {
+  getTestPlanReportById
+} = require('../models/services/TestPlanReportService');
+const {
+  getAtVersionWithRequirements
+} = require('../models/services/AtVersionService');
 
 const scheduleCollectionJobResolver = async (
-  _,
+  _parent,
   { testPlanReportId },
   context
 ) => {
@@ -14,7 +20,27 @@ const scheduleCollectionJobResolver = async (
     throw new AuthenticationError();
   }
 
-  return scheduleCollectionJob({ testPlanReportId }, { transaction });
+  const report = await getTestPlanReportById({
+    id: testPlanReportId,
+    transaction
+  });
+
+  if (!report) {
+    throw new Error(`Test plan report with id ${testPlanReportId} not found`);
+  }
+
+  // Get the appropriate AT version for automation
+  const atVersion = await getAtVersionWithRequirements(
+    report.at.id,
+    report.exactAtVersion,
+    report.minimumAtVersion,
+    transaction
+  );
+
+  return scheduleCollectionJob(
+    { testPlanReportId, atVersion },
+    { transaction }
+  );
 };
 
 module.exports = scheduleCollectionJobResolver;
