+++ server/seeders/20220516200024-test_results_at_browser_version.js
@@ -4,7 +4,7 @@ const {
   getTestPlanRuns,
   updateTestPlanRunById
 } = require('../models/services/TestPlanRunService');
-const { getAtVersionByQuery } = require('../models/services/AtService');
+const { getAtVersionByQuery } = require('../models/services/AtVersionService');
 const {
   getBrowserVersionByQuery
 } = require('../models/services/BrowserService');
