--- server/util/getAtVersionWithRequirements.js
@@ -1,62 +1,0 @@
-const { Op } = require('sequelize');
-const { getAtVersions } = require('../models/services/AtVersionService');
-const { AT_VERSIONS_SUPPORTED_BY_COLLECTION_JOBS } = require('./constants');
-
-const getAtVersionWithRequirements = async (
-  atId,
-  exactAtVersion,
-  minimumAtVersion,
-  transaction
-) => {
-  try {
-    if (exactAtVersion) {
-      return exactAtVersion;
-    }
-
-    if (!minimumAtVersion) {
-      throw new Error(
-        'Either exactAtVersion or minimumAtVersion must be provided'
-      );
-    }
-
-    if (minimumAtVersion.supportedByAutomation) {
-      return minimumAtVersion;
-    }
-
-    const matchingAtVersions = await getAtVersions({
-      where: {
-        atId,
-        releasedAt: { [Op.gte]: minimumAtVersion.releasedAt }
-      },
-      pagination: {
-        order: [['releasedAt', 'ASC']]
-      },
-      transaction
-    });
-
-    const supportedVersions =
-      AT_VERSIONS_SUPPORTED_BY_COLLECTION_JOBS[
-        matchingAtVersions[0]?.at?.name
-      ] || [];
-
-    const latestSupportedAtVersion = matchingAtVersions.find(
-      atv =>
-        supportedVersions.includes(atv.name) &&
-        new Date(atv.releasedAt) >= new Date(minimumAtVersion.releasedAt)
-    );
-
-    if (!latestSupportedAtVersion) {
-      throw new Error(
-        `No suitable AT version found for automation for AT ${atId} ` +
-          `with minimumAtVersion ${minimumAtVersion?.name}`
-      );
-    }
-
-    return latestSupportedAtVersion;
-  } catch (error) {
-    console.error('Error while determining AT version:', error);
-    throw error;
-  }
-};
-
-module.exports = getAtVersionWithRequirements;
